requirements.txt

Django==5.2.2
Pillow==11.2.1  # For image handling

djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0

django-cors-headers==4.7.0
djoser==2.3.1

social-auth-app-django==5.5.1
social-auth-core==4.7.0

paystack==1.5.0
python-dateutil==2.8.2

gunicorn==23.0.0
whitenoise==6.9.0  # For static files

django-storages 
cloudinary
django-cloudinary-storage
